Created a Player class with tally_score method
Player class will be used to keep track of data that is specific to each
player in the game. This includes the score, wins, dice rolls.

tally_score method is used to completed F03 which scores how many
diamond and gold the player has scored. This is done by accepting the
final_roll, and calculating the sum of gold and diamonds and multiplying
by 100.