Added turn method for the player's turn
Created a turn method in Player class to execute specific functions
relating to each player's turn. At the moment, turn is executed so that
the turn ends only when the player obtains 3 skulls (F08).

I will later add something that will randomize which the player keeps to
finish the method.