Updated Object-oriented design
Introduced an init method to initialize each player to have 0 wins and 0
score. Also made the dice, and rolls specific to each player by making
it an attribute for each player. This eliminates the need to introduce
two roll variables in the main file.

Introduced a rerolls() method to break up turn into components. This
makes it more readable. Also, added a win_update method to remove the
need to do it in the main file.

Also added necessary comments